Move implementations from trades/__init__
