+++ src/backlight/metrics/pl.py
@@ -2,9 +2,10 @@ import pandas as pd
 
 from typing import List
 
-from backlight.datasource import MarketData
+from backlight.datasource.marketdata import MarketData
+from backlight.positions import calc_positions
 from backlight.positions.positions import Positions
-from backlight.trades import Trade, count, flatten, evaluate_pl
+from backlight.trades.trades import Trade, count
 
 
 def _sum(a: pd.Series) -> float:
@@ -26,22 +27,25 @@ def _trade_amount(amount: pd.Series) -> pd.Series:
 
 def calc_trade_peformance(trades: List[Trade], mkt: MarketData) -> pd.DataFrame:
     total_count, win_count, lose_count = count(trades, mkt)
-    trade = flatten(trades)
-    pl = evaluate_pl(trade, mkt)
 
     m = pd.DataFrame.from_records(
         [
-            ("cnt_total", total_count),
+            ("cnt_trade", total_count),
             ("cnt_win", win_count),
             ("cnt_lose", lose_count),
             ("win_ratio", win_count / total_count),
             ("lose_ratio", lose_count / total_count),
-            ("pl", pl),
         ]
     ).set_index(0)
-
     del m.index.name
     m.columns = ["metrics"]
+
+    positions = calc_positions(trades, mkt)
+    m = pd.concat([m, calc_position_performance(positions)], axis=1)
+
+    m.loc[:, "avg_win_pl"] = m["total_win_pl"] / m["cnt_win"]
+    m.loc[:, "avg_lose_pl"] = m["total_lose_pl"] / m["cnt_lose"]
+    m.loc[:, "avg_pl_per_trade"] = m["total_pl"] / m["cnt_trade"]
 
     return m.T
 
@@ -52,10 +56,18 @@ def calc_position_performance(positions: Positions) -> pd.DataFrame:
     trade_amount = _trade_amount(positions.amount)
 
     total_pl = _sum(pl)
+    win_pl = _sum(pl[pl > 0.0])
+    lose_pl = _sum(pl[pl < 0.0])
     average_pl = total_pl / trade_amount
 
     m = pd.DataFrame.from_records(
-        [("avg_pl", average_pl), ("total_pl", total_pl), ("trade_amount", trade_amount)]
+        [
+            ("avg_pl_per_amount", average_pl),
+            ("total_pl", total_pl),
+            ("total_win_pl", win_pl),
+            ("total_lose_pl", lose_pl),
+            ("cnt_amount", trade_amount),
+        ]
     ).set_index(0)
 
     del m.index.name
